Fix "buildpack-select" setting shell variables
The "buildpack-select" function sets "selected_path", but it was not being
persisted in the parent shell, in order to be used from "buildpack-compile".
The function appears to have been run in a subshell as a result of piping the
output to "indent". This changes it to use "process substitution" to pass the
output to "indent" while "buildpack-select" runs in the parent shell.