fix: vendor build directory to fix builds
The upstream heroku/base-images switched to using a RUN mount and removed the inlining of certain files, causing build failures for gliderlabs/herokuish when executing their install scripts. To fix this, we _also_ vendor those files.

Additionally, update the removal of postgresql from 15 to 16 to fix issues with building on arm.