feat: use system bash to avoid extracting bash at runtime
When herokuish executes, it extracts a bash binary into the HOME directory - /app - and then executes as normal. This makes it impossible to run read-only containers, as that extraction happens at both build time (when executing buildpacks) as well as runtime (when running applications).

Rather than attempt to move the bash binary to /tmp, we instead use system bash. While this is a downgrade of bash since we packaged a newer bash, this provides better compatibility with upstream Heroku, which executes shell buildpacks with system bash.